10055 without abs

+++ chapter1/10055.cpp
@@ -6,9 +6,9 @@ int main()
 {
 	freopen("input.txt", "r", stdin);
 
-	unsigned int l, r;
-	while(scanf("%u %u", &l, &r) != EOF)
-		printf("%u\n", (l < r) ? r-l : l-r);
+	long l, r;
+	while(scanf("%ld %ld", &l, &r) != EOF)
+		printf("%ld\n", (l < r) ? r-l : l-r);
 
 	return 0;
 }
